+++ src/generic.rs
@@ -1,6 +1,6 @@
 use rand_core::CryptoRng;
 
-use crate::hazmat::SieveFactory;
+use crate::{hazmat::SieveFactory, Error};
 
 /// Sieves through the results of `sieve_factory` and returns the first item for which `predicate` is `true`.
 ///
@@ -9,7 +9,7 @@ pub fn sieve_and_find<R, S>(
     rng: &mut R,
     sieve_factory: S,
     predicate: impl Fn(&mut R, &S::Item) -> bool,
-) -> Option<S::Item>
+) -> Result<Option<S::Item>, Error>
 where
     S: SieveFactory,
     R: CryptoRng + ?Sized,
@@ -18,16 +18,19 @@ where
     // Unlike the parallel version, it is avoidable here.
 
     let mut sieve_factory = sieve_factory;
-    let mut sieve = sieve_factory.make_sieve(rng, None)?;
+    let mut sieve = match sieve_factory.make_sieve(rng, None)? {
+        Some(sieve) => sieve,
+        None => return Ok(None),
+    };
 
     loop {
         if let Some(value) = sieve.find(|num| predicate(rng, num)) {
-            return Some(value);
+            return Ok(Some(value));
         }
-        if let Some(new_sieve) = sieve_factory.make_sieve(rng, Some(&sieve)) {
+        if let Some(new_sieve) = sieve_factory.make_sieve(rng, Some(&sieve))? {
             sieve = new_sieve;
         } else {
-            return None;
+            return Ok(None);
         }
     }
 }
@@ -37,7 +40,7 @@ mod tests {
     use rand_core::{CryptoRng, OsRng, TryRngCore};
 
     use super::sieve_and_find;
-    use crate::hazmat::SieveFactory;
+    use crate::{hazmat::SieveFactory, Error};
 
     #[test]
     fn test_exhaustable_sieve_factory() {
@@ -54,22 +57,22 @@ mod tests {
                 &mut self,
                 _rng: &mut R,
                 previous_sieve: Option<&Self::Sieve>,
-            ) -> Option<Self::Sieve> {
+            ) -> Result<Option<Self::Sieve>, Error> {
                 self.count += 1;
                 if previous_sieve.is_none() {
-                    Some(self.count * 10..(self.count * 10 + 2))
+                    Ok(Some(self.count * 10..(self.count * 10 + 2)))
                 } else {
-                    None
+                    Ok(None)
                 }
             }
         }
 
         let factory = TestSieveFactory { count: 0 };
         let result = sieve_and_find(&mut OsRng.unwrap_mut(), factory, |_rng, num| *num == 11);
-        assert!(result.is_some());
+        assert!(result.unwrap().is_some());
 
         let factory = TestSieveFactory { count: 0 };
         let result = sieve_and_find(&mut OsRng.unwrap_mut(), factory, |_rng, num| *num == 20);
-        assert!(result.is_none());
+        assert!(result.unwrap().is_none());
     }
 }