+++ src/hazmat/miller_rabin.rs
@@ -298,7 +298,10 @@ mod tests {
     fn trivial() {
         let mut rng = ChaCha8Rng::from_seed(*b"01234567890123456789012345678901");
         let start = random_odd_integer::<U1024, _>(&mut rng, NonZero::new(1024).unwrap(), SetBits::Msb).unwrap();
-        for num in SmallPrimesSieve::new(start.get(), NonZero::new(1024).unwrap(), false).take(10) {
+        for num in SmallPrimesSieve::new(start.get(), NonZero::new(1024).unwrap(), false)
+            .unwrap()
+            .take(10)
+        {
             let mr = MillerRabin::new(Odd::new(num).unwrap());
 
             // Trivial tests, must always be true.